Reorganization
Reorganized code. It now looks nice and pretty and organized. Will work on some better navigation and styling later.
