+++ ReviewEverything/back-end/Movie_Back/dbstart.bat
@@ -1,1 +1,1 @@
-"C:\Program Files\MongoDB\Server\4.0\bin\mongod.exe" --dbpath "C:\Users\Marisollllllll\Desktop\ReviewEverything\ReviewEverything\Peeps"
+"C:\Program Files\MongoDB\Server\4.0\bin\mongod.exe" --dbpath "C:\Users\Austin Blandford\Documents\GitHub\ReviewEverything\ReviewEverything\data"
